+++ src/main/java/ru/khasang/cachoeira/view/createnewprojectwindow/panes/FieldsPane.java
@@ -1,4 +1,4 @@
-package ru.khasang.cachoeira.view.createnewprojectwindow;
+package ru.khasang.cachoeira.view.createnewprojectwindow.panes;
 
 import javafx.geometry.Insets;
 import javafx.geometry.VPos;
